requirements.txt
mlflow==2.10.0
dvc==3.45.0
dvc-s3==3.0.1
scikit-learn==1.4.0
pandas==2.2.0
numpy==1.26.3
tensorflow==2.15.0
torch==2.2.0
xgboost==2.0.3
optuna==3.5.0
pyyaml==6.0.1
boto3==1.34.0
kubernetes==29.0.0
prometheus-client==0.19.0
pytest==7.4.4
pytest-cov==4.1.0
black==24.1.0
flake8==7.0.0
mypy==1.8.0
pydantic==2.5.3
fastapi==0.109.0
uvicorn==0.27.0
celery==5.3.4
redis==5.0.1
sqlalchemy==2.0.25
alembic==1.13.1
wandb==0.16.2
tensorboard==2.15.1